refactor(primitives): move ScrapeData to scrape module
